+++ app/src/main/java/app/doggy/mybrary/ui/book/register/RegisterViewModel.kt
@@ -2,7 +2,7 @@ package app.doggy.mybrary.ui.book.register
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.doggy.core.common.util.UnixTime
+import app.doggy.mybrary.core.common.util.UnixTime
 import app.doggy.mybrary.R
 import app.doggy.mybrary.core.domain.model.author.Author
 import app.doggy.mybrary.core.domain.model.author.AuthorId
