--- core/common/src/main/java/app/doggy/core/common/coroutines/dispatcher/Dispatcher.kt
@@ -1,7 +1,0 @@
-package app.doggy.core.common.coroutines.dispatcher
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.RUNTIME)
-annotation class Dispatcher(val mybraryDispatcher: MybraryDispatchers)
