--- core/common/src/main/java/app/doggy/core/common/coroutines/dispatcher/DispatcherModule.kt
@@ -1,21 +1,0 @@
-package app.doggy.core.common.coroutines.dispatcher
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal object DispatcherModule {
-
-  @Provides
-  @Dispatcher(MybraryDispatchers.DEFAULT)
-  fun providesDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
-
-  @Provides
-  @Dispatcher(MybraryDispatchers.IO)
-  fun providesIoDispatcher(): CoroutineDispatcher = Dispatchers.IO
-}
