--- core/common/src/main/java/app/doggy/core/common/coroutines/dispatcher/MybraryDispatchers.kt
@@ -1,6 +1,0 @@
-package app.doggy.core.common.coroutines.dispatcher
-
-enum class MybraryDispatchers {
-  DEFAULT,
-  IO,
-}
