--- core/common/src/main/java/app/doggy/core/common/exception/NetworkException.kt
@@ -1,6 +1,0 @@
-package app.doggy.core.common.exception
-
-class NetworkException(
-  val code: Int,
-  val errorBody: String,
-) : Throwable()
