--- core/common/src/main/java/app/doggy/core/common/util/UnixTime.kt
@@ -1,4 +1,0 @@
-package app.doggy.core.common.util
-
-@JvmInline
-value class UnixTime(val value: Long)
