+++ core/data/src/main/java/app/doggy/mybrary/core/data/impl/BookRepositoryImpl.kt
@@ -1,16 +1,16 @@
 package app.doggy.mybrary.core.data.impl
 
-import app.doggy.core.common.coroutines.dispatcher.Dispatcher
-import app.doggy.core.common.coroutines.dispatcher.MybraryDispatchers
-import app.doggy.core.database.dao.AuthorDao
-import app.doggy.core.database.dao.BookDao
-import app.doggy.core.database.entity.toBookWithAuthors
-import app.doggy.core.database.util.RoomTransactionExecutor
-import app.doggy.core.network.api.book.BookApi
+import app.doggy.mybrary.core.common.coroutines.dispatcher.Dispatcher
+import app.doggy.mybrary.core.common.coroutines.dispatcher.MybraryDispatchers
+import app.doggy.mybrary.core.database.dao.AuthorDao
+import app.doggy.mybrary.core.database.dao.BookDao
+import app.doggy.mybrary.core.database.entity.toBookWithAuthors
+import app.doggy.mybrary.core.database.util.RoomTransactionExecutor
 import app.doggy.mybrary.core.domain.model.book.Book
 import app.doggy.mybrary.core.domain.model.book.BookId
 import app.doggy.mybrary.core.domain.model.record.Record
 import app.doggy.mybrary.core.domain.repository.BookRepository
+import app.doggy.mybrary.core.network.api.book.BookApi
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
