--- core/network/src/main/java/app/doggy/core/network/api/book/BookApiModule.kt
@@ -1,14 +1,0 @@
-package app.doggy.core.network.api.book
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface BookApiModule {
-
-  @Binds
-  fun bindBookApi(impl: BookApiImpl): BookApi
-}
