--- core/network/src/main/java/app/doggy/core/network/api/book/response/GetVolumesResponse.kt
@@ -1,5 +1,0 @@
-package app.doggy.core.network.api.book.response
-
-internal class GetVolumesResponse(
-  val items: List<ItemResponse>?,
-)
