--- core/network/src/main/java/app/doggy/core/network/api/book/response/ImageLinksResponse.kt
@@ -1,5 +1,0 @@
-package app.doggy.core.network.api.book.response
-
-internal class ImageLinksResponse(
-  val thumbnail: String,
-)
