--- core/network/src/main/java/app/doggy/core/network/api/book/response/VolumeInfoResponse.kt
@@ -1,9 +1,0 @@
-package app.doggy.core.network.api.book.response
-
-internal class VolumeInfoResponse(
-  val title: String,
-  val authors: List<String>?,
-  val description: String?,
-  val pageCount: Int?,
-  val imageLinks: ImageLinksResponse?,
-)
