--- core/network/src/main/java/app/doggy/core/network/retrofit/service/BookService.kt
@@ -1,16 +1,0 @@
-package app.doggy.core.network.retrofit.service
-
-import app.doggy.core.network.api.book.response.GetVolumesResponse
-import retrofit2.Response
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-internal interface BookService {
-
-  @GET("volumes")
-  suspend fun getVolumes(
-    @Query("q") q: String,
-    @Query("maxResults") maxResults: Int,
-    @Query("startIndex") startIndex: Int,
-  ): Response<GetVolumesResponse>
-}
